requirements.txt
Flask==2.2.2
pdfplumber==0.11.5
pytesseract
Pillow==10.4.0
Werkzeug==2.2.2
pandas==2.1.4
openpyxl==3.1.5
gunicorn==20.1.0
